TrackedOp: specify queue sizes and warnings on a per-tracker basis
If we have multiple trackers in a daemon, we want to be able to configure
them separately. Plus, users already know how to control op sizes in the
OSD, so changing the config options (as we did in 2c833c0e21fa19018aba3e2aa791554a955339e3)
is not really appropriate. Instead, provider setters which can be called
at construction time (or on any other change) and use them in the OSD with
the configurables we had previously. Add an observer so you can continue
to change them at run-time

This reverts 2c833c0e21fa19018aba3e2aa791554a955339e3

Signed-off-by: Greg Farnum <greg@inktank.com>