test: Change interpreter from /bin/bash to /bin/sh
Change interpreter from /bin/bash to /bin/sh.  bash is not guaranteed
to be installed on all Unix systems, and it's not guaranteed to be
installed into /bin either.

There are other scripts that specify /bin/bash; they need to be
examined one by one to look for bashisms.  This was the only one I had
to modify to get unit tests working.

Signed-off-by: Alan Somers <asomers@gmail.com>
Signed-off-by: Noah Watkins <noahwatkins@gmail.com>