6748: rgw: Optionally return the bucket name in a response header.
This can be useful in situations where accounting of traffic is done externally
when for example HTTP traffic is cached by a reverse proxy like Varnish.

Since not all traffic reaches the RGW daemon it can't fully account all traffic
and this the caching proxy needs to be aware of which bucket the request came for.

Signed-off-by: Wido den Hollander <wido@42on.com>