osdc/ObjectCacher: back off less during flush
In 36cdd22e35e5aed9931b457f81c4f15cea0878f9 we added a limit to avoid
holding the lock for too long.  However, if we back off, we currently
wait for a full second, which is probably a bit much--we really just want
to give other threads a chance.

Backport: emperor
Signed-off-by: Sage Weil <sage@inktank.com>