organizationmap: match authors with organizations
Using the same format as .mailmap, match author names with the
organization sponsoring them, if any. It can be used from the command
line to display git log statistics with results aggregated by company
names.

The git-check-mailmap command that was introduced in git 1.8.4 can be
used to use .mailmap first and then .organizationmap using the
normalized author names. For instance:

    git log --pretty='%aN <%aE>' | \
    git -c mailmap.file=.organizationmap check-mailmap --stdin

This file is also meant to be a source of information for the maintainers of
http://metrics.ceph.com/

Signed-off-by: Loic Dachary <loic@dachary.org>