mon: remove format argument from osd crush dump
The --format argument of the ceph cli is used to send the desired format
argument. The format argument is always part of the command sent to the
server. Adding it to the command description in MonCommand is not
necessary.

partially revert b713ffbafaec4b2be370fdbfb4bc0849177e7763
revert 6ee8093533bb0126f8762b151e3f5f193a35c15f

Signed-off-by: Loic Dachary <loic@dachary.org>