osd: fix map advance limit to handle map gaps
The recent change in 8099a1353bba5ff5b5b12188236accea52a385c4 would stop
advancing after some number of epochs, but did not take into consideration
the possibilty that there are missing maps.  In that case, it is impossible
to advance past the gap.

Fix this by increasing the max epoch as we go so that we can always get
beyond the gap.

Signed-off-by: Sage Weil <sage@inktank.com>