Revert "Merge pull request #2129 from ceph/wip-librbd-oc"
This reverts commit 728427f220bdf91ffc34a3efe1624648db168f48, reversing
changes made to 9dd90d95831fc6509147227ca0f925d9637b6855.

The dirty_or_tx list is used by flush_set, which means we can
resubmit new IOs for writes that are already in progress.  This
has a compounding effect that overwhelms the OSDs with dup IOs
and stalls out the client.

See, for example, teh failues in this run:
  /a/sage-2014-07-25_17:14:20-fs-wip-msgr-testing-basic-plana

The fix is probably pretty simple, but reverting for now to make
the tests pass.

Signed-off-by: Sage Weil <sage@inktank.com>