PG::init: clear rollback info for backfill as well
Otherwise, we won't remove the old rollback objects from a resurrected pg.  In
rare cases, this can cause us to get an EEXIST if we happen to reuse the same
rename id on the same object in a subsequent interval.

Fixes: #9293
Related to: 4ad37e279bd52902bda911a4572c4120266ee69e
Backport: firefly
Signed-off-by: Samuel Just <sam.just@inktank.com>